Add httptrace
The actuator requires a HttpTraceRepository bean to configure the httptrace endpoint. Now, we can access /httptrace under the actuator and see details of http requests to our application. Remember to authenticate (because of our previous security setup) with the default username of 'user' and the password generated by spring when starting the application.